route handlers can call other functions on the same router
fixes #181
