+++ lib/brisket.js
@@ -8,6 +8,8 @@ var Brisket = {
 
     RouterBrewery: require("./controlling/RouterBrewery"),
 
+    Router: require("./controlling/Router"),
+
     Routers: require("./controlling/Routers"),
 
     Controller: require("./controlling/Controller"),
