+++ lib/client/ClientRenderingWorkflow.js
@@ -16,153 +16,145 @@ var environmentConfigForAllRequests;
 
 var ClientRenderingWorkflow = {
 
-    createHandlerFrom: function(originalHandler) {
-        return function() {
-            currentRequestId = currentRequestId + 1;
+    execute: function(router, originalHandler, params) {
+        currentRequestId = currentRequestId + 1;
 
-            var clientRequest = ClientRequest.from(window, currentRequestId, environmentConfigForAllRequests);
-            var clientResponse = ClientResponse.from(window);
+        var clientRequest = ClientRequest.from(window, currentRequestId, environmentConfigForAllRequests);
+        var clientResponse = ClientResponse.from(window);
 
-            var router = this;
-            var LayoutForRoute = router.layout;
-            var errorViewMappingForRoute = router.errorViewMapping;
-            var onRouteStart = router.onRouteStart;
-            var onRouteComplete = router.onRouteComplete;
+        var LayoutForRoute = router.layout;
+        var errorViewMappingForRoute = router.errorViewMapping;
+        var onRouteStart = router.onRouteStart;
+        var onRouteComplete = router.onRouteComplete;
 
-            var isFirstPageLoad = !currentLayout;
-            var layoutForRouteIsSameAsCurrent = isFirstPageLoad || currentLayout.isSameTypeAs(LayoutForRoute);
-            var shouldUseCurrentLayout = !isFirstPageLoad && layoutForRouteIsSameAsCurrent;
-            var layout = shouldUseCurrentLayout ? currentLayout : new LayoutForRoute();
-            var layoutDelegate = LayoutDelegate.from(layout, LayoutForRoute);
-            var routeArguments = new Array(arguments.length);
+        var isFirstPageLoad = !currentLayout;
+        var layoutForRouteIsSameAsCurrent = isFirstPageLoad || currentLayout.isSameTypeAs(LayoutForRoute);
+        var shouldUseCurrentLayout = !isFirstPageLoad && layoutForRouteIsSameAsCurrent;
+        var layout = shouldUseCurrentLayout ? currentLayout : new LayoutForRoute();
+        var layoutDelegate = LayoutDelegate.from(layout, LayoutForRoute);
 
-            for (var i = 0; i < routeArguments.length; i++) {
-                routeArguments[i] = arguments[i];
+        layout.setEnvironmentConfig(environmentConfigForAllRequests);
+
+        if (isFirstPageLoad) {
+            currentLayout = layout;
+        }
+
+        if (!layoutForRouteIsSameAsCurrent) {
+            whenCurrentLayoutDataFetched = null;
+        }
+
+        var currentLayoutDataHasBeenFetched = whenCurrentLayoutDataFetched && !whenCurrentLayoutDataFetched.isPending();
+
+        var fetchLayoutData = Promise.method(function() {
+            return layout.fetchData();
+        });
+
+        // If handler returns a view, this function returns a promise of a view
+        // If handler returns a promise of a view, this function returns a promise of a view
+        // and absorbs the state of the returned promise
+        var chooseViewForRoute = Promise.method(function() {
+            var allParams = addExtraParamsTo(params,
+                layoutDelegate,
+                clientRequest,
+                clientResponse
+            );
+
+            return originalHandler.apply(router, allParams);
+        });
+
+        function renderLayout(view) {
+            if (!layout.hasBeenRendered) {
+                layout.reattach();
+                layout.render();
             }
 
-            layout.setEnvironmentConfig(environmentConfigForAllRequests);
+            return view;
+        }
 
-            if (isFirstPageLoad) {
-                currentLayout = layout;
-            }
+        function executeLayoutInstructions(requestId) {
+            return function(view) {
 
-            if (!layoutForRouteIsSameAsCurrent) {
-                whenCurrentLayoutDataFetched = null;
-            }
-
-            var currentLayoutDataHasBeenFetched = whenCurrentLayoutDataFetched && !whenCurrentLayoutDataFetched.isPending();
-
-            var fetchLayoutData = Promise.method(function() {
-                return layout.fetchData();
-            });
-
-            // If handler returns a view, this function returns a promise of a view
-            // If handler returns a promise of a view, this function returns a promise of a view
-            // and absorbs the state of the returned promise
-            var chooseViewForRoute = Promise.method(function() {
-                var allParams = addExtraParamsTo(routeArguments,
-                    layoutDelegate,
-                    clientRequest,
-                    clientResponse
-                );
-
-                return originalHandler.apply(router, allParams);
-            });
-
-            function renderLayout(view) {
-                if (!layout.hasBeenRendered) {
-                    layout.reattach();
-                    layout.render();
-                }
-
-                return view;
-            }
-
-            function executeLayoutInstructions(requestId) {
-                return function(view) {
-
-                    if (isCurrentRequest(requestId)) {
-                        if (!isFirstPageLoad) {
-                            layout.backToNormal();
-                        }
-
-                        layoutDelegate.replayInstructions();
-                        layoutDelegate.stopRecording();
+                if (isCurrentRequest(requestId)) {
+                    if (!isFirstPageLoad) {
+                        layout.backToNormal();
                     }
 
-                    return view;
-                };
-            }
-
-            function enterDOM(view) {
-                if (isFirstPageLoad || !layout.isSameAs(currentLayout)) {
-                    layout.enterDOM();
+                    layoutDelegate.replayInstructions();
+                    layoutDelegate.stopRecording();
                 }
 
                 return view;
+            };
+        }
+
+        function enterDOM(view) {
+            if (isFirstPageLoad || !layout.isSameAs(currentLayout)) {
+                layout.enterDOM();
             }
 
-            function close(requestId) {
-                if (isCurrentRequest(requestId)) {
-                    onRouteComplete(layout, clientRequest, clientResponse);
-                    clientRequest.complete();
-                    ClientRequest.setPreviousRequest(clientRequest);
-                }
+            return view;
+        }
 
-                if (notCurrentRequest(requestId) && !layout.isSameAs(currentLayout)) {
-                    layout.close();
-                    router.close();
-                    return;
-                }
+        function close(requestId) {
+            if (isCurrentRequest(requestId)) {
+                onRouteComplete(layout, clientRequest, clientResponse);
+                clientRequest.complete();
+                ClientRequest.setPreviousRequest(clientRequest);
+            }
 
+            if (notCurrentRequest(requestId) && !layout.isSameAs(currentLayout)) {
+                layout.close();
                 router.close();
-
-                if (layoutForRouteIsSameAsCurrent) {
-                    return;
-                }
-
-                currentLayout.close();
-                currentLayout = layout;
+                return;
             }
 
-            if (!currentLayoutDataHasBeenFetched) {
-                whenCurrentLayoutDataFetched = fetchLayoutData();
+            router.close();
+
+            if (layoutForRouteIsSameAsCurrent) {
+                return;
             }
 
-            return (function(requestId) {
+            currentLayout.close();
+            currentLayout = layout;
+        }
 
-                onRouteStart(layout, clientRequest, clientResponse);
+        if (!currentLayoutDataHasBeenFetched) {
+            whenCurrentLayoutDataFetched = fetchLayoutData();
+        }
 
-                return whenCurrentLayoutDataFetched
-                    .then(
-                        chooseViewForRoute
-                    )
-                    .then(
-                        renderLayout
-                    )
-                    .then(
-                        executeLayoutInstructions(requestId)
-                    )
-                    .then(
-                        renderView(layout, requestId)
-                    )
-                    .then(
-                        enterDOM
-                    )
-                    .caught(
-                        redirectOrLogAndRenderError(layout, errorViewMappingForRoute, clientRequest, requestId)
-                    )
-                    .lastly(function() {
-                        try {
-                            return close(requestId);
-                        } catch (e) {
-                            Errors.notify(e, clientRequest);
-                            throw e;
-                        }
-                    });
+        return (function(requestId) {
 
-            })(currentRequestId);
-        };
+            onRouteStart(layout, clientRequest, clientResponse);
+
+            return whenCurrentLayoutDataFetched
+                .then(
+                    chooseViewForRoute
+                )
+                .then(
+                    renderLayout
+                )
+                .then(
+                    executeLayoutInstructions(requestId)
+                )
+                .then(
+                    renderView(layout, requestId)
+                )
+                .then(
+                    enterDOM
+                )
+                .caught(
+                    redirectOrLogAndRenderError(layout, errorViewMappingForRoute, clientRequest, requestId)
+                )
+                .lastly(function() {
+                    try {
+                        return close(requestId);
+                    } catch (e) {
+                        Errors.notify(e, clientRequest);
+                        throw e;
+                    }
+                });
+
+        })(currentRequestId);
     },
 
     setEnvironmentConfig: function(environmentConfig) {
