+++ lib/controlling/Router.js
@@ -3,6 +3,8 @@
 var Backbone = require("../application/Backbone");
 var noop = require("../util/noop");
 var Promise = require("bluebird");
+var Environment = require("../environment/Environment");
+var _ = require("underscore");
 
 var Router = Backbone.Router.extend({
 
@@ -31,9 +33,78 @@ var Router = Backbone.Router.extend({
         return Promise.reject({
             status: statusCode
         });
+    },
+
+    route: function(route, name, callback) {
+        var rawRoute = route;
+
+        if (!_.isRegExp(route)) {
+            route = this._routeToRegExp(route);
+        }
+
+        if (_.isFunction(name)) {
+            callback = name;
+            name = '';
+        }
+
+        if (!callback) {
+            callback = this[name];
+        }
+
+        if (Environment.isServer()) {
+            addRouteToServerDispatcher(this, route, rawRoute, callback);
+        }
+
+        if (Environment.isClient()) {
+            addRouteToBackboneHistory(this, route, rawRoute, name, callback);
+        }
+
+        return this;
     }
 
 });
+
+function addRouteToServerDispatcher(router, route, rawRoute, callback) {
+    var ServerDispatcher = require("../server/ServerDispatcher");
+    var ServerRenderingWorkflow = require("../server/ServerRenderingWorkflow");
+
+    ServerDispatcher.addRoute(route, rawRoute, function(fragment, expressRequest, environmentConfig, clientAppRequirePath) {
+        var params = router._extractParameters(route, fragment);
+
+        if (!callback) {
+            return;
+        }
+
+        var responseForRoute = ServerRenderingWorkflow.execute(
+            router,
+            callback,
+            params,
+            expressRequest,
+            environmentConfig,
+            clientAppRequirePath
+        );
+
+        return responseForRoute;
+    });
+}
+
+function addRouteToBackboneHistory(router, route, rawRoute, name, callback) {
+    var ClientRenderingWorkflow = require("../client/ClientRenderingWorkflow");
+
+    Backbone.history.route(route, function(fragment) {
+        var params = router._extractParameters(route, fragment);
+
+        if (!callback) {
+            return;
+        }
+
+        ClientRenderingWorkflow.execute(router, callback, params);
+
+        router.trigger.apply(router, ["route:" + name].concat(params));
+        router.trigger("route", name, params);
+        Backbone.history.trigger("route", router, name, params);
+    });
+}
 
 module.exports = Router;
