+++ lib/controlling/RouterBrewery.js
@@ -2,10 +2,6 @@
 
 var _ = require("underscore");
 var Router = require("./Router");
-var ServerRouter = require("../server/ServerRouter");
-var ServerRenderingWorkflow = require("../server/ServerRenderingWorkflow");
-var ClientRenderingWorkflow = require("../client/ClientRenderingWorkflow");
-var Environment = require("../environment/Environment");
 
 var RouterBrewery = {
 
@@ -26,57 +22,9 @@ var RouterBrewery = {
 };
 
 function createRouter(props, classProps) {
-    var RouterType = Environment.isServer() ? ServerRouter : Router;
-    var BrisketRouter = RouterType.extend(augmentRouteHandlers(props), classProps);
-
-    BrisketRouter.extend = throwsNoExtendAllowed;
+    var BrisketRouter = Router.extend(props, classProps);
 
     return BrisketRouter;
-}
-
-function throwsNoExtendAllowed() {
-    throw new Error(
-        "You cannot extend a Router. You can only create it using a brewery. " +
-        "Use either the default Brisket.RouterBrewery.create or create your own brewery " +
-        "with the Brisket.RouterBrewery.makeBreweryWithDefaults function."
-    );
-}
-
-function makeHandlerThatRendersView(originalHandler) {
-    if (Environment.isServer()) {
-        return ServerRenderingWorkflow.createHandlerFrom(originalHandler);
-    }
-
-    if (Environment.isClient()) {
-        return ClientRenderingWorkflow.createHandlerFrom(originalHandler);
-    }
-}
-
-function augmentRouteHandlers(props) {
-    if (!props) {
-        return null;
-    }
-
-    var resultProps = _.clone(props);
-
-    for (var route in props.routes) {
-        if (props.routes.hasOwnProperty(route)) {
-            var handler = props[props.routes[route]];
-
-            if (!handler) {
-                throw new Error(
-                    "Route '" + route + "' does not have a handler"
-                );
-            }
-
-            //handlers are excuted in reverse order
-            handler = makeHandlerThatRendersView(handler);
-
-            resultProps[props.routes[route]] = handler;
-        }
-    }
-
-    return resultProps;
 }
 
 module.exports = RouterBrewery;
