+++ lib/server/ServerRenderingWorkflow.js
@@ -12,90 +12,86 @@ var validateView = require("../util/validateView");
 
 var ServerRenderingWorkflow = {
 
-    createHandlerFrom: function(originalHandler) {
-        return function(params, expressRequest, environmentConfig, clientAppRequirePath) {
-            var serverRequest = ServerRequest.from(expressRequest, environmentConfig);
-            var serverResponse = ServerResponse.create();
+    execute: function(router, originalHandler, params, expressRequest, environmentConfig, clientAppRequirePath) {
+        var serverRequest = ServerRequest.from(expressRequest, environmentConfig);
+        var serverResponse = ServerResponse.create();
 
-            var routeArguments = params;
-            var router = this;
-            var LayoutForRoute = router.layout;
-            var errorViewMapping = router.errorViewMapping;
-            var layout = new LayoutForRoute();
-            var layoutDelegate = LayoutDelegate.from(layout, LayoutForRoute);
+        var LayoutForRoute = router.layout;
+        var errorViewMapping = router.errorViewMapping;
+        var layout = new LayoutForRoute();
+        var layoutDelegate = LayoutDelegate.from(layout, LayoutForRoute);
 
-            layout.setEnvironmentConfig(environmentConfig);
+        layout.setEnvironmentConfig(environmentConfig);
 
-            var fetchLayoutData = Promise.method(function() {
-                return layout.fetchData();
-            });
+        var fetchLayoutData = Promise.method(function() {
+            return layout.fetchData();
+        });
 
-            // If handler returns a view, this function returns a promise of a view
-            // If handler returns a promise of a view, this function returns a promise of a view
-            // and absorbs the state of the returned promise
-            var chooseViewForRoute = Promise.method(function() {
-                var allParams = addExtraParamsTo(routeArguments,
-                    layoutDelegate,
-                    serverRequest,
-                    serverResponse
-                );
+        // If handler returns a view, this function returns a promise of a view
+        // If handler returns a promise of a view, this function returns a promise of a view
+        // and absorbs the state of the returned promise
+        var chooseViewForRoute = Promise.method(function() {
+            var allParams = addExtraParamsTo(params,
+                layoutDelegate,
+                serverRequest,
+                serverResponse
+            );
 
-                return originalHandler.apply(router, allParams);
-            });
+            return originalHandler.apply(router, allParams);
+        });
 
-            function renderLayout(view) {
-                layout.render();
+        function renderLayout(view) {
+            layout.render();
 
-                return view;
-            }
+            return view;
+        }
 
-            function executeLayoutInstructions(view) {
-                layoutDelegate.replayInstructions();
-                layoutDelegate.stopRecording();
+        function executeLayoutInstructions(view) {
+            layoutDelegate.replayInstructions();
+            layoutDelegate.stopRecording();
 
-                return view;
-            }
+            return view;
+        }
 
-            function close() {
-                layout.close();
-                router.close();
-                AjaxCallsForCurrentRequest.clear();
-            }
+        function close() {
+            layout.close();
+            router.close();
+            AjaxCallsForCurrentRequest.clear();
+        }
 
-            var forRoute = {
-                layout: layout,
-                environmentConfig: environmentConfig,
-                clientAppRequirePath: clientAppRequirePath,
-                serverRequest: serverRequest,
-                serverResponse: serverResponse,
-                errorViewMapping: errorViewMapping
-            };
-
-            return fetchLayoutData()
-                .then(
-                    chooseViewForRoute
-                )
-                .then(
-                    renderLayout
-                )
-                .then(
-                    executeLayoutInstructions
-                )
-                .then(
-                    renderPage(forRoute)
-                )
-                .caught(
-                    redirectOrLogAndRenderError(forRoute)
-                )
-                .lastly(function() {
-                    try {
-                        return close();
-                    } catch (e) {
-                        Errors.notify(e, serverRequest);
-                        throw e;
-                    }
-                });
+        var forRoute = {
+            layout: layout,
+            environmentConfig: environmentConfig,
+            clientAppRequirePath: clientAppRequirePath,
+            serverRequest: serverRequest,
+            serverResponse: serverResponse,
+            errorViewMapping: errorViewMapping
         };
+
+        return fetchLayoutData()
+            .then(
+                chooseViewForRoute
+            )
+            .then(
+                renderLayout
+            )
+            .then(
+                executeLayoutInstructions
+            )
+            .then(
+                renderPage(forRoute)
+            )
+            .caught(
+                redirectOrLogAndRenderError(forRoute)
+            )
+            .lastly(function() {
+                try {
+                    return close();
+                } catch (e) {
+                    Errors.notify(e, serverRequest);
+                    throw e;
+                }
+            });
     }
 
 };
