--- spec/server/RouterBreweryOnServerSpec.js
@@ -1,41 +1,0 @@
-"use strict";
-
-describe("RouterBrewery on server", function() {
-    var RouterBrewery = require("../../lib/controlling/RouterBrewery");
-    var Environment = require("../../lib/environment/Environment");
-    var ServerRouter = require("../../lib/server/ServerRouter");
-
-    var ExampleRouter;
-    var router;
-
-    beforeEach(function() {
-        spyOn(Environment, "isServer").and.returnValue(true);
-    });
-
-    it("creates ServerRouters", function() {
-        createRouter();
-
-        expect(router instanceof ServerRouter).toBe(true);
-    });
-
-    function createRouter() {
-        ExampleRouter = RouterBrewery.create();
-        router = new ExampleRouter();
-    }
-
-});
-
-// ----------------------------------------------------------------------------
-// Copyright (C) 2016 Bloomberg Finance L.P.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-// http://www.apache.org/licenses/LICENSE-2.0
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-// ----------------------------- END-OF-FILE ----------------------------------
